% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{add_prefix}
\alias{add_prefix}
\title{Add prefix}
\usage{
add_prefix(str, prefix)
}
\arguments{
\item{str}{A string.}

\item{prefix}{A string.}
}
\value{
A string.
}
\description{
Add prefix
}
\keyword{internal}
