% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{all_attributes_character}
\alias{all_attributes_character}
\title{All attributes are character}
\usage{
all_attributes_character(instances)
}
\arguments{
\item{instances}{A tibble.}
}
\value{
A tibble.
}
\description{
All attributes are character
}
\keyword{internal}
