% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{calculate_inherited_relationships}
\alias{calculate_inherited_relationships}
\title{calculate inherited relationships}
\usage{
calculate_inherited_relationships(gd, level_name = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level_name}{A string, name of the lower level.}
}
\description{
Each level has explicitly defined relationships with other levels. For a
given level, all the relationships with the levels of the dimension, direct
and indirect, are obtained.
}
\keyword{internal}
