% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{check_related_levels}
\alias{check_related_levels}
\alias{check_related_levels.geodimension}
\title{Check related levels in a dimension}
\usage{
check_related_levels(gd, lower_level_name, upper_level_name)

\method{check_related_levels}{geodimension}(gd, lower_level_name = NULL, upper_level_name = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{lower_level_name}{A string, name of the lower level.}

\item{upper_level_name}{A string, name of the upper lever.}
}
\value{
A \code{tibble}, unrelated lower level instances.
}
\description{
Given two previously related levels of a dimension, it obtains the instances
of the lower level that have not been related to the upper level.
}
\examples{
region <-
  geolevel(name = "region",
           layer = layer_us_region,
           key = "geoid")

division <-
  geolevel(name = "division",
           layer = layer_us_division,
           key = "geoid")

gd <-
  geodimension(name = "gd_us",
               level = region) |>
  add_level(division)

gd <- gd |>
  relate_levels(lower_level_name = "division",
                upper_level_name = "region",
                by_geography = TRUE)

t <- gd |>
  check_related_levels(lower_level_name = "division",
                       upper_level_name = "region")

}
\seealso{
Other level association functions: 
\code{\link{complete_relation_by_geography}()},
\code{\link{get_unrelated_instances}()},
\code{\link{relate_levels}()}
}
\concept{level association functions}
