% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{define_relationship}
\alias{define_relationship}
\title{define relationship}
\usage{
define_relationship(gd, gdp, l, h)
}
\arguments{
\item{gd}{A \code{geolevel} object.}

\item{gdp}{A \code{geolevel} object.}

\item{l}{A string, name of the lower level.}

\item{h}{A string, name of the upper lever.}
}
\value{
A \code{geolevel} object.
}
\description{
define relationship
}
\keyword{internal}
