% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gd_es}
\alias{gd_es}
\title{\code{gd_es}}
\format{
A \code{geodimension}.
}
\source{
\url{https://centrodedescargas.cnig.es/CentroDescargas/},
\url{https://www.mapa.gob.es/es/cartografia-y-sig/ide/descargas/agricultura/default.aspx}
}
\usage{
gd_es
}
\description{
\code{geodimension} obtained from vector layers over Spain
}
\details{
It includes the levels nucleus, municipality, agricultural region, province,
autonomous community and country.
}
\examples{
# Defined by:

\donttest{
file <- system.file("extdata", "es_layers.gpkg", package = "geodimension")
layer_es_municipality <-
  sf::st_read(file, layer = "municipality", quiet = TRUE)
layer_es_province <-
  sf::st_read(file, layer = "province", quiet = TRUE)
layer_es_autonomous_community <-
  sf::st_read(file, layer = "autonomous_community", quiet = TRUE)

municipality <-
  geolevel(
    name = "municipality",
    layer = layer_es_municipality,
    attributes = c("COD_PROV", "NAME", "CAPITAL", "REGION"),
    key = "COD_INE"
  ) |>
  add_geometry(layer = coordinates_to_geometry(
    layer_es_municipality,
    lon_lat = c("LON_ETRS89", "LAT_ETRS89"),
    crs = 4258
  ))

agricultural_region <-
  geolevel(
    name = "agricultural_region",
    layer = get_level_layer(municipality),
    attributes = "REGION",
    key = c("COD_PROV", "REGION")
  ) |>
  complete_point_geometry()

province <-
  geolevel(
    name = "province",
    layer = layer_es_province,
    attributes = c("COUNTRY", "COD_CA", "NAME", "CAPITAL_PROVINCE"),
    key = "COD_PROV"
  ) |>
  complete_point_geometry()

autonomous_community <-
  geolevel(
    name = "autonomous_community",
    layer = layer_es_autonomous_community,
    attributes = c("COUNTRY", "NAME"),
    key = "COD_CA"
  ) |>
  complete_point_geometry()

country <-
  geolevel(
    name = "country",
    layer = get_level_layer(autonomous_community),
    attributes = "COUNTRY",
    key = "COUNTRY"
  ) |>
  complete_point_geometry()

gd_es <-
  geodimension(name = "gd_es",
               level = municipality,
               snake_case = TRUE) |>
  add_level(level = agricultural_region) |>
  add_level(level = province) |>
  add_level(level = autonomous_community) |>
  add_level(level = country)

gd_es <- gd_es |>
  relate_levels(
    lower_level_name = "municipality",
    lower_level_attributes = "COD_PROV",
    upper_level_name = "province"
  ) |>
  relate_levels(
    lower_level_name = "municipality",
    lower_level_attributes = c("COD_PROV", "REGION"),
    upper_level_name = "agricultural_region"
  ) |>
  relate_levels(
    lower_level_name = "agricultural_region",
    lower_level_attributes = "COD_PROV",
    upper_level_name = "province"
  ) |>
  relate_levels(
    lower_level_name = "province",
    lower_level_attributes = "COD_CA",
    upper_level_name = "autonomous_community"
  ) |>
  relate_levels(
    lower_level_name = "autonomous_community",
    lower_level_attributes = "country",
    upper_level_name = "country"
  )
}

}
\keyword{datasets}
