% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gd_us_city}
\alias{gd_us_city}
\title{\code{gd_us_city}}
\format{
A \code{geodimension}.
}
\source{
\url{https://www.census.gov}
}
\usage{
gd_us_city
}
\description{
\code{geodimension} obtained from vector layers over USA.
}
\details{
It includes the levels city, county, state, region, division and nation.
}
\keyword{datasets}
