% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_geometry}
\alias{get_geometry}
\title{Get geometry}
\usage{
get_geometry(layer)
}
\arguments{
\item{layer}{A \code{sf} object.}
}
\value{
A string.
}
\description{
Get the geometry of a layer, as it is interpreted to define a \code{geolevel}
object.
}
\details{
It will only be valid if one of the three geometries is interpreted: \emph{point},
\emph{line} or \emph{polygon}.
}
\examples{
geometry <- get_geometry(layer_us_region)

}
\seealso{
Other level definition functions: 
\code{\link{check_key}()},
\code{\link{coordinates_to_geometry}()}
}
\concept{level definition functions}
