% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_get_higher_level_names.R
\name{get_higher_level_names}
\alias{get_higher_level_names}
\alias{get_higher_level_names.geodimension}
\title{Add a level to a dimension}
\usage{
get_higher_level_names(gd, level_name = NULL, indirect_levels = FALSE)

\method{get_higher_level_names}{geodimension}(gd, level_name = NULL, indirect_levels = FALSE)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level_name}{A string.}

\item{indirect_levels}{A boolean.}
}
\value{
A vector of names.
}
\description{
Add a level to a dimension. This level can then be related to other levels of
the dimension.
}
\examples{
library(tidyr)


}
\seealso{


Other information gathering functions: 
\code{\link{get_level_data}()},
\code{\link{get_level_geometries}()},
\code{\link{get_level_layer}()},
\code{\link{get_level_names}()}
}
\concept{information gathering functions}
