% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_get_level.R
\name{get_level_data}
\alias{get_level_data}
\alias{get_level_data.geodimension}
\title{Add a level to a dimension}
\usage{
get_level_data(gd, level_name = NULL, inherited = FALSE)

\method{get_level_data}{geodimension}(gd, level_name = NULL, inherited = FALSE)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level_name}{A string.}

\item{inherited}{A boolean.}
}
\value{
A \code{tibble} object.
}
\description{
Add a level to a dimension. This level can then be related to other levels of
the dimension.
}
\examples{
library(tidyr)


}
\seealso{


Other information gathering functions: 
\code{\link{get_level_geometries}()},
\code{\link{get_level_layer}()},
\code{\link{get_level_names}()}
}
\concept{information gathering functions}
