% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explotation.R
\name{get_level_names}
\alias{get_level_names}
\alias{get_level_names.geodimension}
\title{Get level names}
\usage{
get_level_names(gd)

\method{get_level_names}{geodimension}(gd)
}
\arguments{
\item{gd}{A \code{geodimension} object.}
}
\value{
A vector of names.
}
\description{
Get the names of levels included in the \code{geodimension}.
}
\examples{

ln <- gd_us |>
  get_level_names()

}
\seealso{
\code{\link{geodimension}}, \code{\link{geolevel}}

Other information gathering functions: 
\code{\link{get_higher_level_names}()},
\code{\link{get_level_data_geo}()},
\code{\link{get_level_data}()},
\code{\link{get_level_geometries}()},
\code{\link{get_level_layer}()}
}
\concept{information gathering functions}
