% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel.R
\name{\%>\%}
\alias{\%>\%}
\alias{new_geolevel}
\title{\code{geolevel} S3 class}
\usage{
new_geolevel(name = NULL, layer = NULL, attributes = NULL, key = NULL)
}
\arguments{
\item{name}{A string, level name.}

\item{layer}{A \code{sf} object.}

\item{attributes}{A vector, selected attributes.}

\item{key}{A vector, attributes that compose the key.}
}
\value{
A \code{geolevel} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
