% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layer_es_agricultural_region}
\alias{layer_es_agricultural_region}
\title{\code{layer_es_agricultural_region}}
\format{
A \code{sf} object.
}
\source{
\url{https://www.mapa.gob.es/es/cartografia-y-sig/ide/descargas/agricultura/default.aspx}
}
\usage{
layer_es_agricultural_region
}
\description{
Polygon geometry layer, with data for the Agricultural Regions of Spain.
}
\keyword{datasets}
