% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layer_es_municipality}
\alias{layer_es_municipality}
\title{\code{layer_es_municipality}}
\format{
A \code{sf} object.
}
\source{
\url{https://centrodedescargas.cnig.es/CentroDescargas/}
}
\usage{
layer_es_municipality
}
\description{
Polygon geometry layer, with data for the Municipalities of Spain.
}
\keyword{datasets}
