% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layer_es_nucleus}
\alias{layer_es_nucleus}
\title{\code{layer_es_nucleus}}
\format{
A \code{sf} object.
}
\source{
\url{https://centrodedescargas.cnig.es/CentroDescargas/}
}
\usage{
layer_es_nucleus
}
\description{
Point geometry layer, with data for the Population Nucleus of Spain.
}
\keyword{datasets}
