% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layer_es_province}
\alias{layer_es_province}
\title{\code{layer_es_province}}
\format{
A \code{sf} object.
}
\source{
\url{https://centrodedescargas.cnig.es/CentroDescargas/}
}
\usage{
layer_es_province
}
\description{
Polygon geometry layer, with data for the Provinces of Spain.
}
\keyword{datasets}
