% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layer_us_city}
\alias{layer_us_city}
\title{\code{layer_us_city}}
\format{
A \code{sf} object.
}
\source{
\url{https://www.census.gov}
}
\usage{
layer_us_city
}
\description{
Point geometry layer, with data for US cities.
}
\keyword{datasets}
