% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layer_us_region}
\alias{layer_us_region}
\title{\code{layer_us_region}}
\format{
A \code{sf} object.
}
\source{
\url{https://www.census.gov}
}
\usage{
layer_us_region
}
\description{
Polygon geometry layer, with data for US regions.
}
\keyword{datasets}
