% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layer_us_state}
\alias{layer_us_state}
\title{\code{layer_us_state}}
\format{
A \code{sf} object.
}
\source{
\url{https://www.census.gov}
}
\usage{
layer_us_state
}
\description{
Polygon geometry layer, with data for US states.
}
\keyword{datasets}
