% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{my_to_snake_case}
\alias{my_to_snake_case}
\title{To snake case}
\usage{
my_to_snake_case(str)
}
\arguments{
\item{str}{A string.}
}
\value{
A vector of strings.
}
\description{
To snake case
}
\keyword{internal}
