% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_relate_levels.R
\name{relate_levels}
\alias{relate_levels}
\alias{relate_levels.geodimension}
\title{Relate levels in a dimension}
\usage{
relate_levels(
  gd,
  lower_level_name = NULL,
  lower_level_attributes = NULL,
  upper_level_name = NULL,
  upper_level_key = NULL,
  by_geography = FALSE
)

\method{relate_levels}{geodimension}(
  gd,
  lower_level_name = NULL,
  lower_level_attributes = NULL,
  upper_level_name = NULL,
  upper_level_key = NULL,
  by_geography = FALSE
)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{lower_level_name}{A string, name of the lower level.}

\item{lower_level_attributes}{A vector of attribute names.}

\item{upper_level_name}{A string, name of the upper lever.}

\item{upper_level_key}{A vector of attribute names.}

\item{by_geography}{A boolean.}
}
\value{
A \code{geodimension}.
}
\description{
Relate levels in a dimension
}
\examples{
library(tidyr)


}
\seealso{


Other level association functions: 
\code{\link{add_level}()},
\code{\link{complete_relation_by_geography}()},
\code{\link{geodimension}()},
\code{\link{get_unrelated_instances}()}
}
\concept{level association functions}
