% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{select_levels}
\alias{select_levels}
\alias{select_levels.geodimension}
\title{Select levels}
\usage{
select_levels(gd, level_names = NULL)

\method{select_levels}{geodimension}(gd, level_names = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level_names}{A vector of names.}
}
\value{
A \code{geodimension} object.
}
\description{
Select a subset of the levels of the dimension so that the rest of the levels
no longer belong to it.
}
\examples{

region <-
  geolevel(name = "region",
           layer = layer_us_region,
           key = "geoid")

division <-
  geolevel(name = "division",
           layer = layer_us_division,
           key = "geoid")

gd <-
  geodimension(name = "gd_us",
               level = region) |>
  add_level(division)

gd <- gd |>
  relate_levels(lower_level_name = "division",
                lower_level_attributes = "REGION",
                upper_level_name = "region")

gd_1 <- gd |>
  select_levels(level_names = "region")

gd_2 <- gd |>
  select_levels(level_names = "division")

}
\seealso{
Other configuration functions: 
\code{\link{transform_crs}()}
}
\concept{configuration functions}
