% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel.R
\name{snake_case_geolevel}
\alias{snake_case_geolevel}
\title{snake case geolevel}
\usage{
snake_case_geolevel(gl)
}
\arguments{
\item{gl}{A \code{geolevel} object.}
}
\value{
A \code{geolevel} object.
}
\description{
snake case geolevel
}
\keyword{internal}
