% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{sort_by_number_of_instances}
\alias{sort_by_number_of_instances}
\title{sort by number of instances}
\usage{
sort_by_number_of_instances(gd, level_names, decreasing = FALSE)
}
\arguments{
\item{gd}{A \code{geolevel} object.}

\item{level_names}{A string vector.}

\item{decreasing}{A boolean, decreasing order.}
}
\value{
A string vector.
}
\description{
sort by number of instances
}
\keyword{internal}
