% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel.R
\name{surrogate_key_name}
\alias{surrogate_key_name}
\title{Surrogate key name}
\usage{
surrogate_key_name(name)
}
\arguments{
\item{name}{A string.}
}
\value{
A string.
}
\description{
Surrogate key name
}
\keyword{internal}
