% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{transform_crs}
\alias{transform_crs}
\alias{transform_crs.geodimension}
\title{Transform CRS}
\usage{
transform_crs(gd, crs = NULL)

\method{transform_crs}{geodimension}(gd, crs = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{crs}{A coordinate reference system: integer with the EPSG code, or
character with proj4string.}
}
\value{
A \code{geodimension}.
}
\description{
Transform the CRS of all the layers included in the dimension to the one
indicated.
}
\examples{

place <-
  geolevel(name = "place",
           layer = layer_us_place,
           attributes = c("STATEFP", "county_geoid", "NAME", "type"),
           key = "GEOID")

county <-
  geolevel(
    name = "county",
    layer = layer_us_county,
    attributes = c("STATEFP", "NAME", "type"),
    key = "GEOID"
  ) |>
  add_geometry(coordinates_to_geometry(layer_us_county,
                                       lon_lat = c("INTPTLON", "INTPTLAT")))

gd_us <-
  geodimension(name = "gd_us",
               level = place) |>
  add_level(level = county) |>
  transform_crs(crs = 3857)

}
\seealso{
Other configuration functions: 
\code{\link{select_levels}()}
}
\concept{configuration functions}
