% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel_add_geometry.R
\name{add_geometry}
\alias{add_geometry}
\alias{add_geometry.geolevel}
\title{Add geometry to a level}
\usage{
add_geometry(gl, layer = NULL, key = NULL, level_key = NULL)

\method{add_geometry}{geolevel}(gl, layer = NULL, key = NULL, level_key = NULL)
}
\arguments{
\item{gl}{A \code{geolevel} object.}

\item{layer}{A \code{sf} object.}

\item{key}{A vector of string.}

\item{level_key}{A vector of string.}
}
\value{
A \code{geolevel}.
}
\description{
A level can have several associated geometries (point, polygon or line). Add
the geometry of the layer or replace an existing one of the indicated type.
}
\examples{
library(tidyr)


}
\seealso{


Other geodimension functions: 
\code{\link{add_level}()},
\code{\link{complete_point_geometry}()},
\code{\link{complete_relation_by_geography}()},
\code{\link{geodimension}()},
\code{\link{geolevel}()},
\code{\link{get_empty_geometry_instances}()},
\code{\link{get_level_data}()},
\code{\link{get_level_geometries}()},
\code{\link{get_level_layer}()},
\code{\link{get_level_names}()},
\code{\link{get_unrelated_instances}()},
\code{\link{relate_levels}()},
\code{\link{select_levels}()}
}
\concept{geodimension functions}
