% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel_add_geometry.R
\name{add_geometry}
\alias{add_geometry}
\alias{add_geometry.geolevel}
\title{Add geometry to a level}
\usage{
add_geometry(gl, layer = NULL, key = NULL, level_key = NULL)

\method{add_geometry}{geolevel}(gl, layer = NULL, key = NULL, level_key = NULL)
}
\arguments{
\item{gl}{A \code{geolevel} object.}

\item{layer}{A \code{sf} object.}

\item{key}{A vector of string.}

\item{level_key}{A vector of string.}
}
\value{
A \code{geolevel}.
}
\description{
A level can have several associated geometries (\emph{point}, \emph{polygon} or
\emph{line}). Add the geometry of the layer or replace an existing one of the
indicated type.
}
\examples{
library(tidyr)


}
\seealso{


Other level definition functions: 
\code{\link{check_key}()},
\code{\link{complete_point_geometry}()},
\code{\link{geolevel}()},
\code{\link{get_empty_geometry_instances}()},
\code{\link{get_geometry}()}
}
\concept{level definition functions}
