% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{add_level}
\alias{add_level}
\alias{add_level.geodimension}
\title{Add a level to a dimension}
\usage{
add_level(gd, level)

\method{add_level}{geodimension}(gd, level = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level}{A \code{geolevel}, level to add to the dimension.}
}
\value{
A \code{geodimension}.
}
\description{
Once a level is part of the dimension, it can then be related to other levels
of the dimension.
}
\seealso{
Other geodimension definition functions: 
\code{\link{geodimension}()}
}
\concept{geodimension definition functions}
