% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{add_level}
\alias{add_level}
\alias{add_level.geodimension}
\title{Add a level to a dimension}
\usage{
add_level(gd, level)

\method{add_level}{geodimension}(gd, level = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level}{A \code{geolevel}, level to add to the dimension.}
}
\value{
A \code{geodimension}.
}
\description{
Once a level is part of the dimension, it can then be related to other levels
of the dimension.
}
\examples{

file <- system.file("extdata", "us_layers.gpkg", package = "geodimension")
layer_us_place <- sf::st_read(file, layer = "place", quiet = TRUE)
layer_us_county <- sf::st_read(file, layer = "county", quiet = TRUE)

place <-
  geolevel(name = "place",
           layer = layer_us_place,
           attributes = c("STATEFP", "county_geoid", "NAME", "type"),
           key = "GEOID")

county <-
  geolevel(
    name = "county",
    layer = layer_us_county,
    attributes = c("STATEFP", "NAME", "type"),
    key = "GEOID"
  ) |>
  add_geometry(coordinates_to_geometry(layer_us_county,
                                       lon_lat = c("INTPTLON", "INTPTLAT")))

gd_us <-
  geodimension(name = "gd_us",
               level = place) |>
  add_level(level = county)

}
\seealso{
Other geodimension definition functions: 
\code{\link{geodimension}()}
}
\concept{geodimension definition functions}
