% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{complete_relation_by_geography}
\alias{complete_relation_by_geography}
\alias{complete_relation_by_geography.geodimension}
\title{Complete relation by geography}
\usage{
complete_relation_by_geography(
  gd,
  lower_level_name = NULL,
  upper_level_name = NULL
)

\method{complete_relation_by_geography}{geodimension}(
  gd,
  lower_level_name = NULL,
  upper_level_name = NULL
)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{lower_level_name}{A string, name of the lower level.}

\item{upper_level_name}{A string, name of the upper lever.}
}
\value{
A \code{geodimension} object.
}
\description{
Two levels can be related by attributes or by geography (if the upper level
has polygon-type geometry). Once related, if there are unrelated instances,
we can try to relate those instances using this function, which considers
alternative geographic relationships.
}
\details{
It does not necessarily succeed trying to relate instances.
}
\examples{

region <-
  geolevel(name = "region",
           layer = layer_us_region,
           key = "geoid")

division <-
  geolevel(name = "division",
           layer = layer_us_division,
           key = "geoid")

gd <-
  geodimension(name = "gd_us",
               level = region) |>
  add_level(division)

gd <- gd |>
  relate_levels(lower_level_name = "division",
                lower_level_attributes = "REGION",
                upper_level_name = "region")

gd <- gd |>
  complete_relation_by_geography(lower_level_name = "division",
                                 upper_level_name = "region")

}
\seealso{
Other level association functions: 
\code{\link{check_related_levels}()},
\code{\link{get_unrelated_instances}()},
\code{\link{relate_levels}()}
}
\concept{level association functions}
