% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{coordinates_to_geometry}
\alias{coordinates_to_geometry}
\title{Transform coordinates to point geometry}
\usage{
coordinates_to_geometry(table, lon_lat = c("intptlon", "intptlat"), crs = 4326)
}
\arguments{
\item{table}{A \code{tibble} object.}

\item{lon_lat}{A vector, name of longitude and latitude attributes.}

\item{crs}{A coordinate reference system: integer with the EPSG code, or
character with proj4string.}
}
\value{
A \code{sf} object.
}
\description{
From the coordinates defined in fields such as latitude and longitude, it
returns a layer of points.
}
\details{
If we start from a geographic layer, it initially transforms it into a table.

The CRS of the new layer is indicated. By default, it considers 4326 (WGS84).
}
\examples{

file <- system.file("extdata", "us_layers.gpkg", package = "geodimension")
layer_us_county <- sf::st_read(file, layer = "county", quiet = TRUE)

us_county_point <-
  coordinates_to_geometry(layer_us_county,
                          lon_lat = c("INTPTLON", "INTPTLAT"))

}
\seealso{
Other level definition functions: 
\code{\link{check_key}()},
\code{\link{get_geometry}()}
}
\concept{level definition functions}
