% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{define_relationship}
\alias{define_relationship}
\title{snake case geolevel}
\usage{
define_relationship(gl, gdp, l, h)
}
\arguments{
\item{gl}{A \code{geolevel} object.}

\item{l}{A string, name of the lower level.}

\item{h}{A string, name of the upper lever.}

\item{glp}{A \code{geolevel} object.}
}
\value{
A \code{geolevel} object.
}
\description{
snake case geolevel
}
\keyword{internal}
