% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gd_es}
\alias{gd_es}
\title{\code{gd_es}}
\format{
A \code{geodimension}.
}
\source{
\url{https://centrodedescargas.cnig.es/CentroDescargas/},
\url{https://www.mapa.gob.es/es/cartografia-y-sig/ide/descargas/agricultura/default.aspx}
}
\usage{
gd_es
}
\description{
\code{geodimension} obtained from vector layers over Spain
}
\details{
It includes the levels nucleus, municipality, agricultural region, province,
autonomous community and country.
}
\keyword{datasets}
