% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{geodimension}
\alias{geodimension}
\title{\code{geodimension} S3 class}
\usage{
geodimension(name = NULL, level = NULL, snake_case = FALSE)
}
\arguments{
\item{name}{A string, name of the dimension.}

\item{level}{A \code{geolevel}.}

\item{snake_case}{A boolean, transform all names to snake_case.}
}
\value{
A \code{geodimension} object.
}
\description{
A \code{geodimension} object is created. A \code{geodimension} allows us to relate
levels. In addition to the name of the \code{geodimension} , a \code{level} has to be
given.
}
\seealso{
Other geodimension definition functions: 
\code{\link{add_level}()}
}
\concept{geodimension definition functions}
