% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel_add_geometry.R
\name{get_empty_geometry_instances}
\alias{get_empty_geometry_instances}
\alias{get_empty_geometry_instances.geolevel}
\title{get empty geometry}
\usage{
get_empty_geometry_instances(gl, geometry = NULL)

\method{get_empty_geometry_instances}{geolevel}(gl, geometry = NULL)
}
\arguments{
\item{gl}{A \code{geolevel} object.}

\item{geometry}{A string, type of geometry of the layer.}
}
\value{
A \code{tibble}.
}
\description{
A level can have several associated geometries (point, polygon or line). Add
the geometry of the layer or replace an existing one of the indicated type.
}
\examples{
library(tidyr)


}
\seealso{


Other geodimension functions: 
\code{\link{add_geometry}()},
\code{\link{add_level}()},
\code{\link{complete_point_geometry}()},
\code{\link{complete_relation_by_geography}()},
\code{\link{geodimension}()},
\code{\link{geolevel}()},
\code{\link{get_level_data}()},
\code{\link{get_level_geometries}()},
\code{\link{get_level_layer}()},
\code{\link{get_level_names}()},
\code{\link{get_unrelated_instances}()},
\code{\link{relate_levels}()},
\code{\link{select_levels}()}
}
\concept{geodimension functions}
