% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolevel_add_geometry.R
\name{get_empty_geometry_instances}
\alias{get_empty_geometry_instances}
\alias{get_empty_geometry_instances.geolevel}
\title{get empty geometry}
\usage{
get_empty_geometry_instances(gl, geometry = NULL)

\method{get_empty_geometry_instances}{geolevel}(gl, geometry = NULL)
}
\arguments{
\item{gl}{A \code{geolevel} object.}

\item{geometry}{A string, type of geometry of the layer.}
}
\value{
A \code{tibble}.
}
\description{
A level can have several associated geometries (point, polygon or line). Add
the geometry of the layer or replace an existing one of the indicated type.
}
\examples{
library(tidyr)


}
\seealso{


Other level definition functions: 
\code{\link{add_geometry}()},
\code{\link{check_key}()},
\code{\link{complete_point_geometry}()},
\code{\link{geolevel}()},
\code{\link{get_geometry}()}
}
\concept{level definition functions}
