% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension.R
\name{get_higher_level_names}
\alias{get_higher_level_names}
\alias{get_higher_level_names.geodimension}
\title{Get higher level names}
\usage{
get_higher_level_names(gd, level_name, indirect_levels)

\method{get_higher_level_names}{geodimension}(gd, level_name = NULL, indirect_levels = FALSE)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level_name}{A string.}

\item{indirect_levels}{A boolean.}
}
\value{
A vector of names.
}
\description{
Get the names of levels included in the \code{geodimension} that are related to the
given level and are upper levels. We can get only the direct levels or the
levels reached by passing through other levels.
}
\examples{

region <-
  geolevel(name = "region",
           layer = layer_us_region,
           key = "geoid")

division <-
  geolevel(name = "division",
           layer = layer_us_division,
           key = "geoid")

gd <-
  geodimension(name = "gd_us",
               level = region) |>
  add_level(division)

gd <- gd |>
  relate_levels(lower_level_name = "division",
                lower_level_attributes = "REGION",
                upper_level_name = "region")

ln <- gd |>
  get_higher_level_names(level_name = "division",
                         indirect_levels = TRUE)

}
\seealso{
Other information output functions: 
\code{\link{get_level_data}()},
\code{\link{get_level_geometries}()},
\code{\link{get_level_layer}()},
\code{\link{get_level_names}()}
}
\concept{information output functions}
