% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_get_level.R
\name{get_level_geometries}
\alias{get_level_geometries}
\alias{get_level_geometries.geodimension}
\title{Add a level to a dimension}
\usage{
get_level_geometries(gd, level_name = NULL)

\method{get_level_geometries}{geodimension}(gd, level_name = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level_name}{A string.}
}
\value{
A vector of names.
}
\description{
Add a level to a dimension. This level can then be related to other levels of
the dimension.
}
\examples{
library(tidyr)


}
\seealso{


Other geodimension functions: 
\code{\link{add_geometry}()},
\code{\link{add_level}()},
\code{\link{complete_point_geometry}()},
\code{\link{complete_relation_by_geography}()},
\code{\link{geodimension}()},
\code{\link{geolevel}()},
\code{\link{get_empty_geometry_instances}()},
\code{\link{get_level_data}()},
\code{\link{get_level_layer}()},
\code{\link{get_level_names}()},
\code{\link{get_unrelated_instances}()},
\code{\link{relate_levels}()}
}
\concept{geodimension functions}
