% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_get_level_geometries.R
\name{get_level_geometries}
\alias{get_level_geometries}
\alias{get_level_geometries.geodimension}
\title{Get level geometries}
\usage{
get_level_geometries(gd, level_name = NULL)

\method{get_level_geometries}{geodimension}(gd, level_name = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level_name}{A string.}
}
\value{
A vector of names.
}
\description{
Gets the geometry types defined for a given level.
}
\examples{
library(tidyr)

lg <- gd_us \%>\%
  get_level_geometries(level_name = "state")

}
\seealso{


Other information output functions: 
\code{\link{get_higher_level_names}()},
\code{\link{get_level_data}()},
\code{\link{get_level_layer}()},
\code{\link{get_level_names}()}
}
\concept{information output functions}
