% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_get_level.R
\name{get_level_layer}
\alias{get_level_layer}
\alias{get_level_layer.geodimension}
\title{Add a level to a dimension}
\usage{
get_level_layer(
  gd,
  level_name = NULL,
  attributes = FALSE,
  surrogate_key = FALSE,
  inherited = FALSE,
  geometry = NULL
)

\method{get_level_layer}{geodimension}(
  gd,
  level_name = NULL,
  attributes = FALSE,
  surrogate_key = FALSE,
  inherited = FALSE,
  geometry = NULL
)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level_name}{A string.}

\item{attributes}{A boolean.}

\item{surrogate_key}{A boolean.}

\item{inherited}{A boolean.}

\item{geometry}{A string.}
}
\value{
A \code{sf} object.
}
\description{
Add a level to a dimension. This level can then be related to other levels of
the dimension.
}
\examples{
library(tidyr)


}
\seealso{


Other information gathering functions: 
\code{\link{get_higher_level_names}()},
\code{\link{get_level_data}()},
\code{\link{get_level_geometries}()},
\code{\link{get_level_names}()}
}
\concept{information gathering functions}
