% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_get_level_names.R
\name{get_level_names}
\alias{get_level_names}
\alias{get_level_names.geodimension}
\title{Add a level to a dimension}
\usage{
get_level_names(gd)

\method{get_level_names}{geodimension}(gd)
}
\arguments{
\item{gd}{A \code{geodimension} object.}
}
\value{
A vector of names.
}
\description{
Add a level to a dimension. This level can then be related to other levels of
the dimension.
}
\examples{
library(tidyr)


}
\seealso{


Other information gathering functions: 
\code{\link{get_higher_level_names}()},
\code{\link{get_level_data}()},
\code{\link{get_level_geometries}()},
\code{\link{get_level_layer}()}
}
\concept{information gathering functions}
