% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{get_unrelated_instances}
\alias{get_unrelated_instances}
\alias{get_unrelated_instances.geodimension}
\title{Check related levels in a dimension}
\usage{
get_unrelated_instances(gd, lower_level_name, upper_level_name)

\method{get_unrelated_instances}{geodimension}(gd, lower_level_name = NULL, upper_level_name = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{lower_level_name}{A string, name of the lower level.}

\item{upper_level_name}{A string, name of the upper lever.}
}
\value{
A \code{tibble}, unrelated lower level instances.
}
\description{
Given two previously related levels of a dimension, it obtains the instances
of the lower level that have not been related to the upper level.
}
\seealso{
Other level association functions: 
\code{\link{complete_relation_by_geography}()},
\code{\link{relate_levels}()}
}
\concept{level association functions}
