% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layer_us_county}
\alias{layer_us_county}
\title{\code{layer_us_county}}
\format{
A \code{sf} object.
}
\source{
\url{https://www.census.gov}
}
\usage{
layer_us_county
}
\description{
Polygon geometry layer, with data for US counties.
}
\keyword{datasets}
