% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_select_levels.R
\name{select_levels}
\alias{select_levels}
\alias{select_levels.geodimension}
\title{Select levels}
\usage{
select_levels(gd, level_names = NULL)

\method{select_levels}{geodimension}(gd, level_names = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{level_names}{A vector of names.}
}
\value{
A \code{geodimension} object.
}
\description{
Select a subset of the levels of the dimension so that the rest of the levels
no longer belong to it.
}
\examples{
library(tidyr)

gds <- gd_us \%>\%
  select_levels(level_names = c("division", "region", "nation"))

}
\seealso{


Other configuration functions: 
\code{\link{transform_crs}()}
}
\concept{configuration functions}
