% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_configuration.R
\name{transform_crs}
\alias{transform_crs}
\alias{transform_crs.geodimension}
\title{Add a level to a dimension}
\usage{
transform_crs(gd, crs = NULL)

\method{transform_crs}{geodimension}(gd, crs = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{crs}{A coordinate reference system: integer with the EPSG code, or
character with proj4string.}
}
\value{
A \code{geodimension}.
}
\description{
Add a level to a dimension. This level can then be related to other levels of
the dimension.
}
\examples{
library(tidyr)


}
\seealso{


Other configuration functions: 
\code{\link{select_levels}()}
}
\concept{configuration functions}
