% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodimension_transform_crs.R
\name{transform_crs}
\alias{transform_crs}
\alias{transform_crs.geodimension}
\title{Transform CRS}
\usage{
transform_crs(gd, crs = NULL)

\method{transform_crs}{geodimension}(gd, crs = NULL)
}
\arguments{
\item{gd}{A \code{geodimension} object.}

\item{crs}{A coordinate reference system: integer with the EPSG code, or
character with proj4string.}
}
\value{
A \code{geodimension}.
}
\description{
Transform the CRS of all the layers included in the dimension to the one
indicated.
}
\seealso{


Other configuration functions: 
\code{\link{select_levels}()}
}
\concept{configuration functions}
