% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_division}
\alias{us_division}
\title{\code{us_division}}
\format{
A \code{tibble} object.
}
\usage{
us_division
}
\description{
US Divisions and Regions (code and name).
}
\keyword{datasets}
